pyproject.toml
[project]
name = "rbac"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
	"fastapi>=0.116.0",
	"uvicorn>=0.35.0",
	"sqlalchemy>=2.0.40",
	"python-jose[cryptography]>=3.3.0",
	"passlib[bcrypt]>=1.7.4",
	"python-multipart>=0.0.20",
	"pydantic>=2.11.0",
]
